dashboard/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
jinja2==3.1.3
python-multipart==0.0.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
